# static/swagger.yaml
openapi: 3.0.3
info:
  title: Trading Algorithms
  description:
  version: 1.0.11

tags:
- name: Auth
  description: Authentication
- name: Configuration
  description: Get possible configurations
- name: Trading Algorithms
  description: Run simulations and benchmarks

paths:
  /:
    get:
      summary: Test
      description: Test
      operationId: test
      requestBody:
      responses:
        "200":
          description: Successful operation
        "401":
          description: Invalid username or password
  /auth:
    post:
      tags:
      - Auth
      summary: Get bearer token
      description: Get bearer token
      operationId: auth
      requestBody:
        description: Add the username and password
        content:
          application/json:
            schema:
              required:
              - username
              - password
              type: object
              properties:
                username:
                  type: string
                  example: test
                password:
                  type: string
                  example: test
              example:
                username: test
                password: test
        required: true
      responses:
        "200":
          description: Successful operation
        "401":
          description: Invalid username or password
      security:
      - bearerAuth: []
  /configuration:
    get:
      tags:
        - Configuration
      summary: Get possible configurations
      description: Get possible configurations
      operationId: configuration
      responses:
        "200":
          description: Successful operation
      security:
        - bearerAuth: []
  /simulate:
    post:
      tags:
        - Trading Algorithms
      summary: Run a simulation with the desired configuration
      description: Run a simulation with the desired configuration
      operationId: simulate
      requestBody:
        description: Add the simulation settings
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Mean_Reversion'
                - $ref: '#/components/schemas/Double_RSI'
                - $ref: '#/components/schemas/Arbitrage'
        required: true
      responses:
        "200":
          description: Successful operation
        "401":
          description: Invalid username or password
      security:
        - bearerAuth: [ ]
components:
  schemas:
    Trading_Algorithm:
      type: object
      required:
        - ticker
        - period
        - interval
      properties:
        ticker: 
          type: string
          example: AAPL
        period:
          type: string
          example: 12mo
        interval: 
          type: string
          example: 1d

    Mean_Reversion:
      allOf:
       - $ref: '#/components/schemas/Trading_Algorithm'
       - type: object
         properties:
          algorithm: 
            type: string
            enum: ["mean_reversion"]
          time_window:
            type: integer
            default: 20

    Double_RSI:
      allOf:
       - $ref: '#/components/schemas/Trading_Algorithm'
       - type: object
         properties:
            algorithm: 
              type: string
              enum: ["double_rsi"]
            rsi_short_period:
              type: integer
              default: 14
            rsi_long_period:
              type: integer
              default: 28

    Arbitrage:
      allOf:
        - $ref: '#/components/schemas/Trading_Algorithm'
        - type: object
          properties:
            algorithm: 
              type: string
              enum: ["arbitrage"]
            ticker2:
              type: string
            entry_threshold:
              type: integer
              default: 2
            exit_threshold:
              type: integer
              default: 0

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT